// frontend/src/types.ts
export interface RegisterFormData {
  email: string;
  username: string;
  password: string;
  confirmPassword: string;
  phone: string;
}
export interface IUser {
  _id?: string; // Optional - will be auto-generated by MongoDB
 username: string;
  email?: string;
  password?: string;
  phone?: string;
  createdAt?: Date;
  updatedAt?: Date;
}
export interface IBlog {
  _id?: string;                
  title: string;              
  content: string;
  category:string;           
  image: string;              
  userId?: IUser|string;               
  createdAt?: Date;           // Optional: auto-generated by Mongoose
  updatedAt?: Date;           // Optional: auto-generated by Mongoose
}